requirements.txt
Flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
Faker==20.1.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
openpyxl==3.1.5
PyJWT==2.8.0
Flask-Limiter==3.5.0
cryptography==42.0.5
bcrypt==4.1.2

